# Conflicts:
#	src/main/java/com/utn/space/venueaapi/controllers/ReservationController.java
#	src/main/java/com/utn/space/venueaapi/model/records/ReservationDTO.java
#	src/main/java/com/utn/space/venueaapi/service/ReservationService.java

+++ src/main/java/com/utn/space/venueaapi/controllers/NotificationController.java
@@ -23,14 +23,14 @@ public class NotificationController {
     }
 
     @GetMapping("/consumer/{id}")
-    public ResponseEntity<List<Notification>>findAllByIdConsumer(@PathVariable Long id){
+    public ResponseEntity<List<Notification>>findAllByIdConsumer(@PathVariable Integer id){
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(notificationService.listAllByIdConsumer(id));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Notification> findForId (@PathVariable Long id){
+    public ResponseEntity<Notification> findForId (@PathVariable Integer id){
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(notificationService.findById(id));
@@ -38,7 +38,7 @@ public class NotificationController {
 
 
     @PostMapping("/{id}")
-    public ResponseEntity<Notification> markAsSeen (@PathVariable Long id){
+    public ResponseEntity<Notification> markAsSeen (@PathVariable Integer id){
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(notificationService.markAsSeen(id));
